+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 List new features, changes, and bug fixes by version
+## [0.2.1](https://github.com/jamesjtb/uesrpg-companion/releases/tag/v0.2.1)
+### Changes:
+- Added Initiative Total to Combatant Listing if combat is running
+  - Clicking the Initiative Total for a Combatant Listing will allow the user to edit the combatant's initiative total.
+- Add combat stop, set all initiative totals to null
 ## [0.2.0](https://github.com/jamesjtb/uesrpg-companion/releases/tag/v0.2.0)
 ### New features:
 - Display the current version in the top bar title